183: Dont use separate folder for dashboard footer, PR requested change
