+++ frontend/src/components/DashboardLayout.tsx
@@ -1,7 +1,7 @@
 import Sidebar from './sidebar/Sidebar.tsx';
 import { Outlet } from 'react-router-dom';
 import Navbar from './navbar/Navbar.tsx';
-import DashboardFooter from './dashboardFooter/DashboardFooter.tsx';
+import DashboardFooter from './footer/DashboardFooter';
 import { Box, Flex } from '@mantine/core';
 import { ScrollToTop } from '../components/ScrollToTop';
