--- frontend/src/components/dashboardFooter/DashboardFooterCentered.module.css
@@ -1,23 +1,0 @@
-.footer {
-  margin-top: auto;
-  /* This should be set to auto to keep footer at the bottom of the page*/
-  border-top: rem(1px) solid light-dark(var(--mantine-color-gray-2), var(--mantine-color-dark-5));
-}
-
-.inner {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: var(--mantine-spacing-md) var(--mantine-spacing-md);
-
-  @media (max-width: 768) {
-    flex-direction: column;
-  }
-}
-
-.links {
-  @media (max-width: 768) {
-    margin-top: var(--mantine-spacing-lg);
-    margin-bottom: var(--mantine-spacing-sm);
-  }
-}
